requirements.txt
selenium
webdriver-manager
beautifulsoup4
requests
gspread
google-auth
PyYAML
pandas
